requirements.txt
humanize
lameenc
phonenumbers
salt-pepper